ui: Add searchbar on subproject-level
The searchbar takes and inputstring and use it as Regex to
search each Column for all subproject titles and status.
Unmatched columns are hidden, matched columns are shown
and the matched strings are highlighted.
The searchbar has placeholder-texts in every language,
e.g. "Search projects" or "Search subprojects"
The searchbar on project-level is adjusted to have one
component for two searching jobs.

Resolves: #422